projects/twitter/requirements.txt
numpy
pandas
python-dotenv
tweepy[async]
black
jupyterlab
networkx
pyvis
matplotlib
